refactor os-explode packages
Split the main.go file into separate files in separate packages to
make the program more easily extendable and modifiable in the future
