+++ pkg/doc.go
@@ -1,0 +1,2 @@
+// Package pkg contains all of the libraries for os-explode
+package pkg
